third

+++ PackageFile/Test.cpp
@@ -21,13 +21,13 @@ void main()
 	pPackage->Seek(0, SoPackageFile::Seek_End, theFile);
 	__int64 nFileSize = 0;
 	pPackage->Tell(nFileSize, theFile);
-	char* pBuff = (char*)malloc(nFileSize);
+	char* pBuff = (char*)malloc((size_t)nFileSize);
 	pPackage->Seek(0, SoPackageFile::Seek_Set, theFile);
 	__int64 nActuallyReadCount = 0;
 	pPackage->Read(pBuff, 1, nFileSize, nActuallyReadCount, theFile);
 	pPackage->Close(theFile);
 	FILE* pFile = fopen("D:/DestFile.ddd", "w+b");
-	fwrite(pBuff, 1, nFileSize, pFile);
+	fwrite(pBuff, 1, (size_t)nFileSize, pFile);
 	fclose(pFile);
 	free(pBuff);
 	delete pPackage;
